+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Improvements to timestamps and editing
   - icon colors for different states (not editing, editing, edited)
   - Offset active one above the timeline
+- Hovering over a timestamp in the sidebar will raise it's icon on the timeline up
 
 ## `v1.0.20-beta4`